Merged revision 811 from the UnboundID repository. ------------------------------------------------------------------------ r811 | boli | 2011-12-15 12:56:20 -0600 (Thu, 15 Dec 2011) | 3 lines
Introduces a BasicAuthenticationFilter that basically injects the authenticated username into the request if authentication succeeds.  The Wink SCIM resource implementations will make sure there is an authID in the SecurityContext before proceeding with the rest of the processing. This also fixes an undiscovered issue were 401 responses from SCIM did not contain a SCIM error response in the body. 

JiraIssue: DS-176
------------------------------------------------------------------------