Merged revision 865 from the UnboundID repository. ------------------------------------------------------------------------ r865 | coulbeck | 2012-01-09 12:10:25 -0600 (Mon, 09 Jan 2012) | 16 lines
===BEGIN-RELEASE-NOTE===
Fix an issue with the scim-query-rate tool when run with the --numIntervals option.  If any of the threads created by the tool were blocked waiting for a server response then the tool would also block after the prescribed number of intervals and would not release any of the connections it had been using.
===END-RELEASE-NOTE===

===BEGIN-INTERNAL-COMMENT===
There will now be an error message for each blocked thread when the tool completes, caused when all the connections are terminated. For example:

21559 [QueryRate Thread 8] INFO org.apache.http.impl.client.DefaultHttpClient - I/O exception (java.net.SocketException) caught when processing request: Socket closed
21559 [QueryRate Thread 8] INFO org.apache.http.impl.client.DefaultHttpClient - Retrying request

===END-INTERNAL-COMMENT===

JiraIssue:SCIM-225

Product:scim-ri

------------------------------------------------------------------------