Merged revision 897 from the UnboundID repository. ------------------------------------------------------------------------ r897 | coulbeck | 2012-01-30 10:20:24 -0600 (Mon, 30 Jan 2012) | 12 lines
There was a bug in the JSON parser. In content that contains a single resource, and also in the Bulk request (where there are resources of mixed type), the schemas element of each resource specifies which schemas are used. However, in the query response content, where there are multiple resources of the same type, the schemas are specified once for all resources rather than for each resource.

Also, I have added a toString() method on a resource object and made some toString() methods less verbose and hence far more useful for debugging.

===BEGIN-RELEASE-NOTE===
Fix an issue where extension attributes in a JSON query response were being discarded during client parsing of the response.
===END-RELEASE-NOTE===

JiraIssue:SCIM-247

Product:scim-sdk

------------------------------------------------------------------------