Merged revision 905 from the UnboundID repository. ------------------------------------------------------------------------ r905 | coulbeck | 2012-02-07 10:32:19 -0600 (Tue, 07 Feb 2012) | 13 lines
===BEGIN-RELEASE-NOTE===
Update the SCIM resource configuration to allow the SCIM resource ID to be mapped to an LDAP attribute as an alternative to mapping to the LDAP entry DN. The DN does not necessarily meet the requirements of the SCIM specification regarding resource ID immutability since LDAP permits entries to be renamed or moved. The entryUUID attribute, whose read-only value is assigned by the Directory Server, is a possible alternative mapping.
===END-RELEASE-NOTE===

===BEGIN-INTERNAL-COMMENT===
<optional comments to QA etc. here>
===END-INTERNAL-COMMENT===

JiraIssue:SCIM-149

Product:scim-ri
Product:scim-extension

------------------------------------------------------------------------