Merged revision 919 from the UnboundID repository. ------------------------------------------------------------------------ r919 | tthompson | 2012-02-13 11:54:18 -0600 (Mon, 13 Feb 2012) | 10 lines
===BEGIN-RELEASE-NOTE===
Provide a new PreemptiveAuthInterceptor class which can be used to configure the SCIMService to use preemptive authentication. This reduces the overhead of making requests over an authenticated connection.
===END-RELEASE-NOTE===

This was previously a private class in the ScimQueryRate tool, but we are making it public so that we can use it elsewhere, such as in the unit tests and in the SCIMSyncDestination.

JiraIssue:SCIM-189

Product:scim-sdk

------------------------------------------------------------------------