Merged revision 1109 from the UnboundID repository. ------------------------------------------------------------------------ r1109 | boli | 2012-04-30 14:21:28 -0500 (Mon, 30 Apr 2012) | 4 lines
Change UserResource.isActive() to return a Boolean instead of a preemptive to allow a null return value to indicate a value is not present in the resource. Javadoc is also updated to indicate that null is a possible return value. Added unit-test to make sure all UserResource.get methods that return a simple value will not throw an unexpected exception.

JiraIssue: SCIM-300
Product: scim
------------------------------------------------------------------------