Merged revision 1705 from the UnboundID repository. ------------------------------------------------------------------------ r1705 | tthompson | 2012-09-06 16:32:10 -0500 (Thu, 06 Sep 2012) | 16 lines
===BEGIN-RELEASE-NOTE===
Add support for the PATCH operation in the SCIM SDK and server implementation. This allows for resources to be partially updated without having to send the entire contents of the resource across the network, reducing network and processing overhead. This is especially beneficial for resources with many attributes, such as Groups.
===END-RELEASE-NOTE===

===BEGIN-INTERNAL-COMMENT===
This commit also addresses some of the schema changes with the SCIM 1.1 specification. The remainder of those will be addressed with DS-6892.
Also, there are no new changes to the SCIM Extension itself, as it now lives under the core server, but there were new unit tests added  to
the SCIMExtensionTestCase and these will be ported to the core server in a later commit.
===END-INTERNAL-COMMENT===

JiraIssue: DS-6851
JiraIssue: DS-6853

Product: scim-sdk
Product: scim-extension

------------------------------------------------------------------------