Merged revision 1778 from the UnboundID repository. ------------------------------------------------------------------------ r1778 | tthompson | 2012-09-18 14:22:00 -0500 (Tue, 18 Sep 2012) | 13 lines
===BEGIN-RELEASE-NOTE===
Add support for authentication to SCIM via OAuth 2.0 bearer tokens. This requires an OAuthTokenHandler extension built with the UnboundID Server SDK in order to decode and validate tokens.
===END-RELEASE-NOTE===

===BEGIN-INTERNAL-COMMENT===
This commit adds the required SCIM changes in order to support the OAuthTokenHandler extension from the core server. There will be a corresponding commit in the core server which adds the Server SDK API and server-side glue. Also, support for bulk PATCH operations is flushed out here as well.
===END-INTERNAL-COMMENT===

JiraIssue: DS-6763

Product: scim-sdk
Product: scim-extension

------------------------------------------------------------------------