Merged revision 2197 from the UnboundID repository. ------------------------------------------------------------------------ r2197 | tthompson | 2013-01-24 14:04:43 -0600 (Thu, 24 Jan 2013) | 7 lines
Update the SCIM LDAPBackend so that it respects the "count" parameter when searching across multiple base DNs. It cannot perfectly handle sorting and pagination when the search spans multiple base DNs (because a VLV index is per-backend), but this fix does make sure that exactly the correct number of results are returned, based on the "maxResults" configuration setting and the "count" parameter from the incoming request.

Produt: scim-sdk
Product: scim-ri

JiraIssue: SCIM-151

------------------------------------------------------------------------