Merged revision 2598 from the UnboundID repository. ------------------------------------------------------------------------ r2598 | tthompson | 2013-05-14 15:43:50 -0500 (Tue, 14 May 2013) | 7 lines
Revert several of the changes from r2584 because there have been a couple significant issues with the 4.3-beta1 release of the Apache HttpClient. Version 4.2.5 was released a few weeks after 4.3-beta1 and has the fixes we need for DS-8517. Since 4.2.5 is a stable GA, this commit updates SCIM to depend on that version.

JiraIssue: DS-8517

Product: scim-sdk
Product: scim-ri

------------------------------------------------------------------------