Merged revision 2758 from the UnboundID repository. ------------------------------------------------------------------------ r2758 | lsaenz | 2013-07-26 15:09:53 -0500 (Fri, 26 Jul 2013) | 20 lines
We went back to the SCIM board after reading some ambiguous implementation docs about how this should be handled. The consensus seems to be that readOnly attributes should NEVER be set via SCIM, and are only to be handled server side.
So the new behavior is this:
POST - iterate through the attributes on the presented ScimObject, and fail if any are read only.
PUT - iterate through the attributes on the presented ScimObject, and remove all read only attrs before processing differences to create modifications list
PATCH - make it work like POST

Unfortunately this is left our customer in a bad state, since there is no idea of "set only at create time", which is what they were trying to accomplish. So I decided to only enforce the POST failure for CORE schema. Which means that you can set a read only custom schema attribute at POST time, but you will never again be able to update it via SCIM. 

===BEGIN-RELEASE-NOTE===
Fix a bug that allowed SCIM requests to update attributes that were designated as 'read only' in the SCIM Resource Schema.
===END-RELEASE-NOTE===

Reviewer: Andy Coulbeck, Trevor Thompson

JiraIssue:DS-9068

SalesforceCaseNumber:1962

Product:ds

------------------------------------------------------------------------