Merged revision 2800 from the UnboundID repository. ------------------------------------------------------------------------ r2800 | lsaenz | 2013-08-14 09:56:22 -0500 (Wed, 14 Aug 2013) | 11 lines
Change the way we handle PUT for read only attributes again to prevent unintended deletion of read only attrs.
The original fix just removed the read only attrs from the request object since that's the way the spec is written, but in practice this does not work because the modification detection thinks that you are trying to delete all of the missing read only attrs.  So now I'm using my original code that determines which mods are due to read only attr changes, and removes those mods.

Reviewer: Trevor Thompson

JiraIssue:DS-9068

SalesforceCaseNumber:1962

Product:ds

------------------------------------------------------------------------