Merged revision 3436 from the UnboundID repository. ------------------------------------------------------------------------ r3436 | lsaenz | 2014-03-25 14:44:20 -0500 (Tue, 25 Mar 2014) | 15 lines
These changes add an option for implicit schema declaration
which works similarly to the older versions of SCIM.  The older versions of SCIM unmarshalled JSON data by looking at each key and treating it as either a schema or an attribute name.  But this could lead to attribute descriptions being created for non-existent schemas, which would run into bugs later.
The new method is schema based and assumes each key is a schema, or is an attribute that belongs to the core schema.  It also requires that the core schema be listed in the "schemas".
The option I've added is enabled by setting the system property "com.unboundid.scim.implicit-schema-checking" to true.  This will result in two significant changes.  One, always consider all known schemas from the resource descriptor in addition to the list of provided schemas at unmarshall time.  Two, if the key provided in the JSON stream is not in the list of schemas, then we will look for any schemas with a attribute matching that name, with the core schema winning in the case of multiple matches... otherwise the first match.  We will then use that schema to build the scim object for that key.
This is not intended to be used in production going forward, but as a crutch to provide backwards compatibility for upgrading customers.

Reviewer: Andy Coulbeck

JiraIssue:DS-10421

SalesforceCaseNumber:2169

Product:ds
Product:proxy

------------------------------------------------------------------------