Merged revision 3483 from the UnboundID repository. ------------------------------------------------------------------------ r3483 | boli | 2014-04-18 13:00:53 -0500 (Fri, 18 Apr 2014) | 7 lines
Changed the ResourceMapper in SCIM LDAP so it will only include an attribute descriptor in the schema if there is a mapping or derivation definition. If not, it will issue an warning to the debug logger. Enable debug logging when using the InMemroyServerTool so the user will be warned about unmapped attribute definitions. Commented out all attribute definitions without default mappings in the OOTB resource.xml files. 

Updated unit-tests to play nice with the updated behavior.

JiraIssue: DS-10542
Product: scim-sdk
Reviewer: jchildress
------------------------------------------------------------------------