Merged revision 4280 from the UnboundID repository. ------------------------------------------------------------------------ r4280 | lsaenz | 2015-03-04 09:53:57 -0600 (Wed, 04 Mar 2015) | 10 lines
Updated the LDAP Backend test code to allow for testing without VLV / Simple Paged Results controls. This change adds a test to ensure that simple paged results work as intended. Also added logic to the control processing to get the estimated count from the simple paged results control if the VLV reponse control is not available. Also changed the VLV control processing to look for multiple VLV response controls. We will not log an error when this happens, and we will try not to use a VLV response with zero matches if possible. The results are still unpredictable, but should be more stable.

Reviewer: Bo Li

JiraIssue:DS-12485

SalesforceCaseNumber:2636

Product:ds

------------------------------------------------------------------------